skills/unitygen/tools/requirements.txt
python-dotenv
requests
numpy
pillow
scipy
rembg
pymatting
onnxruntime
